make group_spawn use group().spawn()
