+++ src/stdlib/windows.rs
@@ -1,26 +1,10 @@
 use std::{
-	io::Result,
 	os::windows::{io::AsRawHandle, process::CommandExt},
 	process::Command,
 };
 use winapi::um::winbase::CREATE_SUSPENDED;
 
-use crate::{builder::CommandGroupBuilder, winres::*, CommandGroup, GroupChild};
-
-impl CommandGroup for Command {
-	fn group_spawn(&mut self) -> Result<GroupChild> {
-		let (job, completion_port) = job_object(true)?;
-		self.creation_flags(CREATE_SUSPENDED);
-		let child = self.spawn()?;
-		assign_child(child.as_raw_handle(), job)?;
-
-		Ok(GroupChild::new(child, job, completion_port))
-	}
-
-	fn group<'a>(&'a mut self) -> CommandGroupBuilder<'a, Command> {
-		CommandGroupBuilder::new(self)
-	}
-}
+use crate::{builder::CommandGroupBuilder, winres::*, GroupChild};
 
 impl CommandGroupBuilder<'_, Command> {
 	/// Executes the command as a child process group, returning a handle to it.
