+++ src/tokio/windows.rs
@@ -1,30 +1,7 @@
-use std::io::Result;
 use tokio::process::Command;
 use winapi::um::winbase::CREATE_SUSPENDED;
 
-use crate::{builder::CommandGroupBuilder, winres::*, AsyncCommandGroup, AsyncGroupChild};
-
-#[async_trait::async_trait]
-impl AsyncCommandGroup for Command {
-	fn group_spawn(&mut self) -> Result<AsyncGroupChild> {
-		let (job, completion_port) = job_object(true)?;
-		self.creation_flags(CREATE_SUSPENDED);
-
-		let child = self.spawn()?;
-		assign_child(
-			child
-				.raw_handle()
-				.expect("child has exited but it has not even started"),
-			job,
-		)?;
-
-		Ok(AsyncGroupChild::new(child, job, completion_port))
-	}
-
-	fn group<'a>(&'a mut self) -> CommandGroupBuilder<'a, Command> {
-		CommandGroupBuilder::new(self)
-	}
-}
+use crate::{builder::CommandGroupBuilder, winres::*, AsyncGroupChild};
 
 impl CommandGroupBuilder<'_, Command> {
 	/// Executes the command as a child process group, returning a handle to it.
@@ -38,7 +15,7 @@ impl CommandGroupBuilder<'_, Command> {
 	/// Basic usage:
 	///
 	/// ```no_run
-	/// use std::process::Command;
+	/// use tokio::process::Command;
 	/// use command_group::CommandGroup;
 	///
 	/// Command::new("ls")
@@ -47,12 +24,10 @@ impl CommandGroupBuilder<'_, Command> {
 	///         .expect("ls command failed to start");
 	/// ```
 	pub fn spawn(&mut self) -> std::io::Result<AsyncGroupChild> {
+		let (job, completion_port) = job_object(self.kill_on_drop)?;
 		self.command
 			.creation_flags(self.creation_flags | CREATE_SUSPENDED);
 
-		// note: same a as in AsyncCommandGroup::group_spawn
-		// but without creation_flags(CREATE_SUSPENDED)
-		let (job, completion_port) = job_object(self.kill_on_drop)?;
 		let child = self.command.spawn()?;
 		assign_child(
 			child
